requirements.txt
psycopg2-binary
pandas
plotly
streamlit
groq
python-dotenv
